NodeViewHolder merged into ErgoNodeViewHolder
